Some security methods implemented, code beauty improved

+++ src/main/java/com/innowise/employeeserviceee/service/impl/DepartmentServiceImpl.java
@@ -4,8 +4,8 @@ package com.innowise.employeeserviceee.service.impl;
 import com.innowise.employeeserviceee.dto.DepartmentDTO;
 import com.innowise.employeeserviceee.dto.converter.DepartmentConverter;
 import com.innowise.employeeserviceee.entity.Department;
+import com.innowise.employeeserviceee.exception.NoSuchRecordException;
 import com.innowise.employeeserviceee.repository.DepartmentRepository;
-import com.innowise.employeeserviceee.repository.impl.DepartmentRepositoryImpl;
 import com.innowise.employeeserviceee.service.DepartmentService;
 import jakarta.ejb.EJB;
 import jakarta.ejb.Stateless;
@@ -14,6 +14,7 @@ import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Stateless
@@ -24,7 +25,6 @@ public class DepartmentServiceImpl implements DepartmentService {
     private DepartmentRepository departmentRepository;
     @EJB
     private DepartmentConverter converter;
-
 
 
     @Override
@@ -42,21 +42,21 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public DepartmentDTO findById(Long id) {
-        Department department = departmentRepository.findById(id)
-//                .orElseThrow(() ->
-//                        new NoSuchRecordException(String.format("Department with id=%s not found", id))
-//                )
-                ;
+        Department department = departmentRepository.findById(id);
+
+        if (Objects.isNull(department)) {
+            throw new NoSuchRecordException(String.format("Employee with id=%s not found", id));
+        }
         return converter.toDTO(department);
     }
 
     @Override
     @Transactional
     public void deleteById(Long id) {
-//        if (!departmentRepository.existsById(id)) {
-//            throw new NoSuchRecordException
-//                    (String.format("Department with id=%s not found for deleting", id));
-//        }
+        if (!departmentRepository.existsById(id)) {
+            throw new NoSuchRecordException
+                    (String.format("Department with id=%s not found for deleting", id));
+        }
         departmentRepository.deleteById(id);
     }
 }
